+++ driver/mina/src/test/java/com/starlight/intrepid/StubIntrepidDriver.java
@@ -68,7 +68,7 @@ public class StubIntrepidDriver implements IntrepidDriver {
 
 
 	@Override
-	public void init( InetAddress server_address, Integer server_port, String vmid_hint,
+	public void init( SocketAddress server_address, String vmid_hint,
 		InboundMessageHandler message_handler, ConnectionListener connection_listener,
 		ScheduledExecutor thread_pool, VMID vmid,
 		ThreadLocal<VMID> deserialization_context_vmid,